gatgu/requirements.txt
asgiref==3.2.10
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
boto3==1.17.88
botocore==1.20.88
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
distro==1.5.0
Django==3.1
django-cors-headers==3.7.0
django-debug-toolbar==2.2
django-redis==4.12.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.7.1
docker==5.0.0
docker-compose==1.29.0
dockerpty==0.4.1
docopt==0.6.2
firebase-admin==5.0.1
flake8==3.9.2
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jmespath==0.10.0
jsonschema==3.2.0
mccabe==0.6.1
mock==4.0.3
msgpack==1.0.2
mysqlclient==2.0.1
nose==1.3.7
packaging==21.0
paramiko==2.7.2
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
websocket-client==0.58.0
zope.interface==5.3.0
